Move instance HasClient UVerb to a better place.
