+++ servant-client/servant-client.cabal
@@ -39,7 +39,6 @@ source-repository head
 library
   exposed-modules:
     Servant.Client
-    Servant.Client.UVerb
     Servant.Client.Streaming
     Servant.Client.Internal.HttpClient
     Servant.Client.Internal.HttpClient.Streaming
@@ -78,13 +77,11 @@ library
     , kan-extensions        >= 5.2      && < 5.3
     , monad-control         >= 1.0.2.3  && < 1.1
     , semigroupoids         >= 5.3.1    && < 5.4
-    , sop-core              >= 0.4.0.0  && < 0.6
     , transformers-base     >= 0.4.5.2  && < 0.5
     , transformers-compat   >= 0.6.2    && < 0.7
 
   hs-source-dirs: src
   default-language: Haskell2010
-
   ghc-options: -Wall -Wno-redundant-constraints
 
 test-suite spec
