+++ servant-client/src/Servant/Client.hs
@@ -12,9 +12,7 @@ module Servant.Client
   , defaultMakeClientRequest
   , hoistClient
   , module Servant.Client.Core.Reexport
-  , module Servant.Client.UVerb
   ) where
 
 import           Servant.Client.Core.Reexport
 import           Servant.Client.Internal.HttpClient
-import           Servant.Client.UVerb
